polygons with simple meshes
