+++ Objects/Objects/GIS/GisPolygonGeometry.cs
@@ -1,25 +1,11 @@
-using System;
 using System.Collections.Generic;
-using System.Runtime.CompilerServices;
-using System.Xml.Linq;
+using Objects.Geometry;
 using Speckle.Core.Models;
 
 namespace Objects.GIS;
 
-public class GisFeature : Base
+public class GisPolygonGeometry : Base
 {
-  public GisFeature() { }
-
-  public GisFeature(List<Base> geometry, Base attributes)
-  {
-    this.geometry = geometry;
-    this.attributes = attributes;
-  }
-
-  [DetachProperty]
-  public List<Base>? geometry { get; set; }
-
-  [DetachProperty]
-  public List<Base>? displayValue { get; set; }
-  public Base? attributes { get; set; }
+  public Polyline boundary { get; set; }
+  public List<Polyline>? voids { get; set; }
 }
